% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_variance.R
\name{bs_iso_skew}
\alias{bs_iso_skew}
\title{Bootstrap ISO SKEW}
\usage{
bs_iso_skew(
  rst_fit,
  nbs = 100,
  seed = NULL,
  sample_residual = c("empirical", "parametric"),
  smooth_bs = TRUE,
  verbose = 1,
  ...
)
}
\description{
Bootstrap ISO SKEW
}
