% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_iso_norm}
\alias{fit_iso_norm}
\title{Isotonic regression with Normal Error}
\usage{
fit_iso_norm(init_pa, init_ai, y, x, z, unimodal, max_steps, tol)
}
\description{
Isotonic regression with Normal Error
}
