% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_iso_skew}
\alias{fit_iso_skew}
\title{Isotonic regression with Skewed Error}
\usage{
fit_iso_skew(pa, ai, y, x, z, unimodal, usez, max_steps, tol)
}
\description{
Isotonic regression with Skewed Error
}
