% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_para_skew_old}
\alias{fit_para_skew_old}
\title{Parametric alpha(x) with Skewed Error}
\usage{
fit_para_skew_old(init_pa, y, x, z, usez, max_steps, tol)
}
\description{
alpha(x) =  alpha * x where alpha > 0
}
