% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_tools.R
\name{get_bs_ci}
\alias{get_bs_ci}
\title{Get bootstrap confidence interval}
\usage{
get_bs_ci(bs, method = c("empirical"), quants = c(0.025, 0.975), est = NULL)
}
\description{
Get bootstrap confidence interval
}
