% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_variance.R
\name{get_sample_kernel}
\alias{get_sample_kernel}
\title{Sample from Epanchenikov}
\usage{
get_sample_kernel(n)
}
\description{
Sample from Epanchenikov
}
\details{
1. Generate iid ğ‘ˆ1,ğ‘ˆ2,ğ‘ˆ3 ~ Uniform(-1,1).
2. If |ğ‘ˆ3|â‰¥|ğ‘ˆ2| and |ğ‘ˆ3|â‰¥|ğ‘ˆ1|, deliver ğ‘ˆ2; otherwise deliver ğ‘ˆ3.
}
