% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_methods.R
\name{predict.ISO}
\alias{predict.ISO}
\title{Prediction from fitted isotonic model}
\usage{
\method{predict}{ISO}(object, new_x = NULL, new_z = NULL, h = -1, ...)
}
\description{
Prediction from fitted isotonic model
}
