% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_methods.R
\name{predict.PARA}
\alias{predict.PARA}
\title{Prediction from fitted parametric model}
\usage{
\method{predict}{PARA}(object, new_x = NULL, new_z = NULL, ...)
}
\description{
Prediction from fitted parametric model
}
