% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_methods.R
\name{sf_bootstrap.SKEWFIT}
\alias{sf_bootstrap.SKEWFIT}
\title{Bootstrap}
\usage{
\method{sf_bootstrap}{SKEWFIT}(
  object,
  nbs = 100,
  seed = NULL,
  verbose = 1,
  pred_x = NULL,
  n_cores = 1,
  ...
)
}
\description{
Bootstrap
}
