% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_methods.R
\name{sf_bs_ci.BOOTSTRAP}
\alias{sf_bs_ci.BOOTSTRAP}
\title{Get confidence interval from bootstrap results}
\usage{
\method{sf_bs_ci}{BOOTSTRAP}(
  object,
  method = c("theoretical", "empirical"),
  quants = c(0.025, 0.975),
  ...
)
}
\description{
Get confidence interval from bootstrap results
}
