% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_fit.R
\name{sf_iso_norm}
\alias{sf_iso_norm}
\title{Isotonic regression with normal error}
\usage{
sf_iso_norm(
  y,
  x,
  z,
  init_pa = NULL,
  max_steps = 1e+05,
  tol = 1e-06,
  unimodel = 0
)
}
\description{
Isotonic regression with normal error
}
