% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_fit.R
\name{sf_para_norm}
\alias{sf_para_norm}
\title{Least square estimator with constraints on the slope of x}
\usage{
sf_para_norm(y, x, z)
}
\description{
Least square estimator with constraints on the slope of x
}
