% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_fit.R
\name{sf_para_skew}
\alias{sf_para_skew}
\title{Parametric model with Skew Normal Errors}
\usage{
sf_para_skew(
  y,
  x,
  z,
  usez = 1,
  init_pa = NULL,
  max_steps = 10000,
  tol = 1e-06,
  ...
)
}
\description{
Parametric model with Skew Normal errors with constraints on the slope of x
}
