% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_methods.R
\name{sf_sample_residual.NORM}
\alias{sf_sample_residual.NORM}
\title{Sample residual from normal model}
\usage{
\method{sf_sample_residual}{NORM}(object, n, method = c("empirical", "parametric"), ...)
}
\description{
Sample residual from normal model
}
