% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_methods.R
\name{sf_sample_residual.SKEW}
\alias{sf_sample_residual.SKEW}
\title{Sample residual from SKEW model}
\usage{
\method{sf_sample_residual}{SKEW}(object, n, method = c("empirical", "parametric"), ...)
}
\description{
Sample residual from SKEW model
}
