% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewfit_fit.R
\name{sf_iso_skew}
\alias{sf_iso_skew}
\title{Isotonic regression with skewed error}
\usage{
sf_iso_skew(
  y,
  x,
  z,
  init_pa = NULL,
  init_ai = NULL,
  max_steps = 1e+05,
  tol = 1e-06,
  unimodel = 0,
  usez = 1
)
}
\description{
Isotonic regression with skewed error
}
